refactor(repeater): better naming for repeaters api related entities
relates-to #22
