+++ packages/repeater/src/api/index.ts
@@ -1,0 +1,3 @@
+import './register';
+
+export { RepeatersManager } from './RepeatersManager';
