+++ packages/repeater/src/api/register.ts
@@ -1,0 +1,5 @@
+import { HttpRepeatersManager } from './HttpRepeatersManager';
+import { RepeatersManager } from './RepeatersManager';
+import { container } from 'tsyringe';
+
+container.register(RepeatersManager, { useClass: HttpRepeatersManager });
